Replace only last occurrence of the title in social menu items
By using str_replace the social menu breaks if you add title to the menu which is the same text as the label. For example using Facebook as label and link title will result in broken menu structure. Switched to preg_replace and only getting the last occurrence fixed it for me. Please test :).